+++ forui/lib/src/widgets/tooltip.dart
@@ -16,9 +16,9 @@ class FTooltipController extends FChangeNotifier {
   late final Animation<double> _fade;
   late final Animation<double> _scale;
 
-  /// Creates a [FTooltipController] with the given [vsync] and animation [duration].
-  FTooltipController({required TickerProvider vsync, duration = const Duration(milliseconds: 100)}) {
-    _animation = AnimationController(vsync: vsync, duration: duration);
+  /// Creates a [FTooltipController] with the given [vsync] and animation [animationDuration].
+  FTooltipController({required TickerProvider vsync, Duration animationDuration = const Duration(milliseconds: 100)}) {
+    _animation = AnimationController(vsync: vsync, duration: animationDuration);
     _fade = _fadeTween.animate(_animation);
     _scale = _scaleTween.animate(_animation);
   }